Change file driver naming expectations
Originally, I thought it'd be a good idea to have Carpal do most of the
work when it comes to parsing the resource address and looking for
files. For example, if you get a request with the param
`/.well-know/webfinger?resource=acct:bob@foobar.com`, you can configure
Carpal with your domain of `foobar.com` and a directory containing the
file `bob`, and that'd be where the JRD is generated from.

This is based on a lot of assumptions - we'd only be serving `acct:`
resources, and each Carpal deployment could only serve resources from a
single domain. Instead of getting fancy, I'm just going to make it so
that you have a directory with files matching the full resource name. In
effect, to serve the `GET
/.well-know/webfinger?resource=acct:bob@foobar.com` request, you'd
point Carpal to a directory containing the file `acct:bob@foobar.com`
and that's where the data would be fetched from.